fix(typescript): Make b64 header optional
The b64 header is currently defined as a required field, so omitting it throws a typescript error. Since its type is defined as "never", there is nothing that can be assigned to it, making it impossible to actually use.
This makes b64 optional, which still prevents assigning to it due to the type.