Support decrypting empty ciphertext
When encrypting an empty string/buffer in GCM mode, the resulting ciphertext is also empty. Trying to then decrypt this using compact decryption then throws an error because it converts the empty string to undefined and then throws an error.
This is probably an edge case, but it is valid to encrypt and decrypt an empty value. This is also not an issue in CBC mode, since the ciphertext of an empty value is a non-empty value.